fixed format fields and moved to utils
