+++ tests/unit/test_utils.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from utils import utils
 import shutil
 import unittest
+from pyspark.sql import SparkSession
 
 
 class UtilsTests(unittest.TestCase):
@@ -12,13 +13,20 @@ class UtilsTests(unittest.TestCase):
     tmp_dir = "tmp-out"
 
     def setUp(self):
-        pass
+        spark = SparkSession.builder \
+            .appName("sfc_data_engineering_csv_to_parquet") \
+            .getOrCreate()
+        self.df = spark.read.csv(self.test_csv_path, header=True)
 
     def tearDown(self):
         try:
             shutil.rmtree(self.tmp_dir)
         except OSError as e:
             pass  # Ignore dir does not exist
+        
+        
+        
+        
 
     def test_generate_s3_dir_date_path(self):
 
@@ -46,6 +54,12 @@ class UtilsTests(unittest.TestCase):
 
         self.assertTrue(Path("tmp-out").is_dir())
         self.assertTrue(Path("tmp-out/_SUCCESS").exists())
+
+    def test_format_date_fields(self):
+        self.assertEqual(self.df.select("date_col").first()[0], "28/11/1993")
+        formatted_df = utils.format_date_fields(self.df)
+        self.assertEqual(str(formatted_df.select(
+            "date_col").first()[0]), "1993-11-28 00:00:00")
 
 
 if __name__ == '__main__':
